+++ CHANGELOG.md
@@ -14,6 +14,18 @@ and `Removed`.
 
 ## [Unreleased]
 
+### Added
+
+- Ability to select Ajour data folder when backing up.
+
+### Fixed
+
+- Windows: Only one instance of Ajour can be launched at a time.
+
+### Changed
+
+- Renamed option `both` to `all` for `backup` CLI command.
+
 ## [1.1.0] - 2021-04-15
 
 ### Added